list_subscriber: add detailed info and edit capability to the list
The list can now produce detailed info about a given subscriber. Some of
these fields are also editable by an NREN-admin. The changeable fields
are:

Contact
	- phone
	- email
Responsible person
	- name
	- email
Comment
State
	a) Suspended
	b) Subscribed
	c) Unsubscribed

The DN-name and the IdP-name is *not* editable. If these needs to be
changed, the subscriber-entry must be deleted and a new created.

Signed-off-by: Henrik Austad <henrik@austad.us>