Revoke_certificates: Apply commonName-wildcard in the right way
We queried for "%" . $eppn . "%" in the search for the common-name of the
currently logged on user. However, if the eppn of user A is the suffix of
the eppn of user B, user A will be able to revoke user B's certificates.

For instance, user A has the eppn test@feide.no and user B the eppn
confusatest@feide.no. The search for the certificates of user A will be
%test@feide.no% which will match %confusatest@feide.no% and hence the
certificates of user B. The new search is "% test@feide.no" which will only
match user A's certificates, due to the way we construct the subject-CN.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>