Do not issue certificates for other than subscribed subscribers
Introduce a new convenience class "Permission". Permission can store, if
the permission for a certain operation is granted or denied along with a
list of reasons.

Introduce a method in Person mayRequestCertificate() which checks all the
attributes, the Confusa entitlement and the subscriber's subscription
status (subscribed, suspended, unsubscribed), that can be called in
process_csr. mayRequestCertificate returns a Permission object containing
yes/no and a reason enumeration, that can be displayed to the user.

CertManager::verifyAttributes is obsoleted by that check and removed from
CertManager.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>