CertManager_Online: Filter download cert-list for the organization name
However unlikely it may be that two individuals from different
organizations have the same common-name (due to the fact that the eppn is part
of the CN) it may still be possible. Hence, additionally to the person's
commonName also check the organization name against the respective O=
information in the displayed certificate.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>