Add-subscriber: Use the attribute O-name of the NREN admin as a hint
When I first used the add_subscriber functionality, I was not sure if I
should enter the name of the O-attribute (i.e. "eduPersonOrgDN",
"schacHomeOrganization") or the value of the O-attribute (i.e.
"dc=uninett,dc=no") into the Attribute-name field.

The commit creates a hint that changes this. To make the hint less generic,
it uses the epodn or schacHomeOrganization as sent by the NREN-admin's
IdP.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>