NREN-admin: Give a clear error-message on duplicate subscriber DN-name
The DN-name of a subscriber must be globally unique, because otherwise
different institutions may map to the same organization names in
certificates, which we don't want. If currently an organization name is
duplicated, insertion fails due to the UNIQUE constraint, but no clear
error message is given.

Added an explicit constraint check for the DN name.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>