CertManager_Online: Use remote-API query heuristics when searching for CNs
Decide based on the search string, if the listing request should be
filtered for organizationName first and then locally matched against the
defined commonName or vice versa. As any heuristics, this is more of a
qualified guess than anything else, but for instance if the commonName
search string is just a wildcard "%" it *does* make sense to first filter
the Comodo-API results for the organizationName.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>