Revoke-certificate: do not automatically wrap the search in wildcards
Because wildcards break user expectations and that is bad. Let the admins
add the wildcards themselves where they need and only in case of a search
in user mode add it automatically in front of the user's eppn. This is
better than just taking the user's constructed CN, because the givenName
may change over time, but the user may still want to revoke her old
certificates.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>