Download certificates: Make inspecting certificates a fancier
Now the certificate subject should not stick out of the content space if it
is long any more.

The verbose certificate shows up directly *below* the
order-number or auth_key it refers to. Furthermore, the verbose inspect
view of the certificate can also be collapsed again without performing any
other action.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>