idp: make sure the authority is set in session before using it.
Even if we have a session, the session does not necessarily have to be
AuthN. If the session is not set, we will not be able to get the
authority and the isValid()-assert will fail.

Since we know that the absence of the authority will mean that the user
is *not* authN, we can safely return false here.

Signed-off-by: Henrik Austad <henrik@austad.us>