confusa_auth: do not bomb if subscriber is not set and the user is NREN-admin
If we cannot find the subscriber in the database, this normally
indicates that the subscriber has not signed up for the service and
users should therefore be barred from entry.

However, NREN-admins also belong to a subscriber-organization, and when
they log in, they have the power to add a subscriber, even if the users
for this organization should not be granted access to the certificates
(i.e. the subscriber is not subscribed).

Signed-off-by: Henrik Austad <henrik@austad.us>