confusa_auth: move SimpleSAMLphp-include into the getAuthManager()
The only place we are going to use the autoloader and
simpleSAMLphp-specific code, is when we are including idp.php, so a
double check for including this is redundant.

This change will move all of this into idp.php, as this is the only file
that actually needs this (and all other cases where the
SimpleSAMLphp-library is needed will require idp.php).

Signed-off-by: Henrik Austad <henrik@austad.us>