Download certificate: Fix download problems with Internet explorer
Note that a no-cache directive is the best way to make download in
Internet-Explorer fail. The new Cacha and pragma directives fix downloading
in IE, but I have no idea why. At the same time, it is the only thing that
worked after trying different things for a rather long time and finally
copying those two lines from some Blog.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>