Adapt config-template and Installer for use of dbconfig-common
If dbconfig-common is used to configure the database on a debian system,
the settings made by dbconfig-common should override Confusa's setting
without making conflicting edits of multiple configuration files necessary.

Thus, the following is done:
- dbconfig-common writes a file called confusa_config.inc.php to the
configuration directory
- if the file is present, its db settings override these in
confusa_config.php via an include by resetting the array variables.
- the installer checks if the file is present, and if so it skips the
configuration of DB settings

This adds some platform specific code to Confusa, which I don't like.
However, note that it only affects parts of the Installer and the
config-file and in a very failsafe way (if dbconfig is not used, there are
no consequences at all). And Debian is also one of the most
important target platforms that Confusa wants to support. The most
important reason was that by only having to edit a single set of files for
all platforms is less error prone than writing a Debian-specific installer
and config-file that would have to reflect every update of Confusa's
installer and config-file.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Signed-off-by: Henrik Austad <henrik@austad.us>