process_csr: search for a matching pubkey in csr_cache, not entire CSR
When we store the public-keys in the database, we require a unique
public-key (the authvar). However, process_csr would search after the
complete CSR, even if the database would restrict on the authvar itself.

This inconsistency will pop up whenever a user uploads two different CSRs,
but based on the same keypair.

This change will alter the search to the auth_key, so that the content of the
CSR is no longer relevant.

Signed-off-by: Henrik Austad <henrik@austad.us>