MailManager: Encode text as quoted-printable and wordwrap it
The encoding as quoted-printable showed up as a necessity when wordwrapping
the text, because PHP is unable to wordwrap UTF-8 characters. Furthermore,
support of quoted-printable encoding allows for economically sending texts
with UTF-8 characters in the future (localized notification mails etc.).

Sadly, a builtin quoted-printable function only comes with PHP 5.3.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>