Attribute-mapping: Dump the attribute value itself, not the Person value
The main reason for that change was that the values of the attributes
themselves can be controlled by the admin, to see which attribute is the
right one for a certain person property. This is particularly
handy, if one for instance wants to see the namespace of the
entitlement attribute, when that one hasn't yet been configured correctly
in config.

Furthermore, it is more failsafe by not requiring Confusa to have decorated
the person fully in order to see the attribute values. Lastly, it matches
the dynamically update attribute value display used in the AJAX hints on
attribute values.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>