key_(up|down)load: terminate execution when not in bypass-mode
The key_upload and key_downlad functionality is intended as debugging
only. Thus, if we are in 'proper' mode, these should not be allowed to
run.

Initially, it was thought that by setting the header to "404 Not
Found", Apache would trigger and stop execution. However, this is not
always the case (if at all), so explicit exit(0) have been added.

At the moment, no know reason for setting other exit-signals are known,
thus only an OK-0 is sent.

Signed-off-by: Henrik Austad <henrik@austad.us>