confusa_auth: do not test for ePPN in person at end of decoratePerson()
The test makes sense, however, when we call Person::setEPPN(), person
will throw criticalAttributeException, resulting in a stall of
execution. In other words, if ePPN is not set, the code will *never*
reach this test. And if the ePPN *is* set, the test won't trigger.

Thus, the code will never result in a new MapNotFoundException()

Signed-off-by: Henrik Austad <henrik@austad.us>