Subscriber: let save() update the current value in the database.
This is a convenient way of updating the values of a Subscriber. By
calling save() after values have been changed, the whole synch-to-db
process becomes a lot simpler to follow.

Ideally, the functions in www/ should *never* talk directly with the
database,  only direct the flows in the backend. This is a step towards
that goal.

Signed-off-by: Henrik Austad <henrik@austad.us>