MailManager::qp_encode: Break strings over 75 characters without wordwrap
If there is no space in a line, the MailManager used to hang itself up in
an endless loop trying to start over again from the position of the last
space (being the beginning of the line). Hence if the position of the last
space in a line is the line's beginning, just break the line.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>