list_subscribers: remove ID-explanation from info-field.
The ID is pretty straight forward, and even if we tell the user about
this being *the* unique ID, we know that we have at least on other
unique element (the IdP-name).

Instead of explaining why we don't want to be anywhere near normalized
and start discussing the nefarious reasons for *not* joining two tables
on varchar, just remove the whole thing.

Signed-off-by: Henrik Austad <henrik@austad.us>