nren_admin: do not call all set-functions in an OR'd if-test
When we do this, PHP will stop the execution in the test-part at first
'true' returned. This means that at most one element can be updated this
way.

Instead, set and store for each element and use the result later. That
way the interpreter is 'tricked' into running all tests, and we can
still test to see if calling save() makes sense.

Signed-off-by: Henrik Austad <henrik@austad.us>