menu: add special formatting for pages that require an AuthN-session
So the users will see that a page will redirect them to the
login-section, add a simple formatting-rule to handle this.

Currently, when the user is not authN, but the page requires authn, the
link will be postfixed with a '*' and made italic. Also, the link is
given a title with message that pressing it it will trigger a redirect.

Signed-off-by: Henrik Austad <henrik@austad.us>