Site-template: Force off quirks mode in IE8
Any content above the doctype declaration that is not XML will turn IE7 and
IE8 into "quirks" mode, in which they use a non-standard rendering and
completely break the layout of Confusa (for instance when PHP echos debug
output). At least for IE8 this behaviour can be forced off by sending a
proprietary meta-attribute.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>