cert_manage_online: do not index keys that do not exist.
In some settings, the returned array from Comodo for certificates is
malformed. We are looking for indexes at a very specific format, and
this may blow up.

This test will just ignore those values and continue with the next in
the list. It will not change the current behavior, but it will remove a
set of NOTICE-messages whenever the keys are 'malformed'.

Signed-off-by: Henrik Austad <henrik@austad.us>