Subscriber: Add support for storing newly created subscribers
Add a "new constructor" (actually just default parameters and if/else
branching, but that's all you get with PHP) which constructs a Subscriber
from scratch and split the save() method up in a update and a insert
procedure, dependant on whether the subscriber already has a DB-ID or not.

Add the preferred language as a member variable, to make the object
complete (for storing).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>