Subscriber: Add 'capi_test' prefix in orgName getter, not setter
Arguably a subscriber should not change DN names, but if the prefix
is added and the subscriber subsequently stored in the DB, it will have the
prefix permanently. If the 'capi_test' switch is turned off later, the
organizations will still be called "TEST UNIVERSITY".

I think it's more error free if we keep the test strings as an overlay,
even though the certificates issued in test-mode will have to be revoked
anyways.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>