table_create: add idp_map table and country to nrens
We need to store the idp-map in the database, and all we need, is the
idp_url used in saml20-idp-remote and a reference to the appropriate
NREN. This way, we can use $session->getIdP() to find the url, and from
there, determine the NREN.

In the NREN-table, we need to store the other element we retrieve via
the maps, namely the country. Since each NREN can only belong to one
country, we store this in the nrens-table (and not in the idp_map).

Thus, we can remove the need for symlinking in the Maps we use to
implement this behavior today.

Signed-off-by: Henrik Austad <henrik@austad.us>