Subscriber: remove stripslashes from constructor
stripslashes will make all prior sanitation, which for instance adds a
slash in front of "'" or '"' pointless and may even lead to the possibility
of SQL injections. If there are not any specific problems with real life
eduPersonOrgDNs or schacHomeOrganizations it is recommended to leave it out
of the code.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>