MDB2Wrapper: change signature of execute()
This will change the signature, however, the default value of the new
parameter will *not* change the current behavior. Since update should
not return any value, the wrapper's update should not force execute() to
fetch the rows from the empty result-set.

This change adds a new parameter to execute (default false) to indicate
if the query is an update and the return from the database can be
ignored.

Signed-off-by: Henrik Austad <henrik@austad.us>