NREN: adapt NREN to the new databaselayout with idp_map
The change is subtle yet extensive. The NREN now takes the IdP-name as
argument in the constructor, and from this it will find the NREN and
decorate it with values.

This way, we do not need any of the maps we previously relied upon,
making the simpleSAMLphp-configuration a lot simpler and the delicate
dependency between the two, lesser.

Signed-off-by: Henrik Austad <henrik@austad.us>