confusa_auth: change decoratePerson to take the idp_name as argument.
With the changed NREN, decoratePerson can accept the name of the IdP and
construct an NREN from this. This also means that we get the name and
country 'for free'.

We pass the idp-name as argument as we handle this differently between
idp.php and bypass.php (and probably other modules later on).

Signed-off-by: Henrik Austad <henrik@austad.us>