Nagios status polling: Use critical_errors table in DB
Use a critical_errors table in the DB into which log-messages above a
configurable log-level will be inserted as they occur. This table can then
be read by the status polling site and errors displayed, if they occured.

Having that information stored in a table, makes it easier to mark errors
as being "resolved" by an admin and it will also scale with the size of the
log.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>