Confusa status reporting: Add shell script for admin to handle crit. errors
The shell-script can operate in a "all-error" mode and in an interactive
mode. In the interactive mode it will ask the admin for every critical
error in the DB what to do. In the "forall"-mode, the admin can tell the
script if all errors should be purged (deleted) or marked as resolved
(update operation on the is_resolved column of the critical_errors).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>