Stylist: Exchange "Download" in notification mail with "Send to <email>"
What the admin wants to do with the notification e-mail, is probably have a
look at how it will look like in the inbox of her users. That's why a Send
to... link has been added, so the admin can try to send the customized mail
template to herself. The download button was not that useful, so it was
sacrificed for the sake of a non-overloaded interface.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>