person: test explicitly for types of admin when doing isAdmin()
When a person is added in the database as admin, but lacks the
necessary entitlement, the admin/normal-mode button will appear, and
some of the links, even if the user does not have the correct
privileges.

Granted, the user cannot *do* anything, but the menu should not show
false info.

Signed-off-by: Henrik Austad <henrik@austad.us>