Process-CSR: Proposed workflow change
The most shocking thing is that the CSR listing has been dropped
altogether. This is hard to get used to, but the user will most probably
not need it because the user will rather be interested in certificates than
CSRs.

Uploading a CSR in the common form will lead the user to an approve page,
where the CSR can be signed or deleted. If the CSR is already in the DB,
there will be a clear hint that this is the case and that it is the old CSR
that the user will sign.

Signing takes place by approving the CSR in the approve mask.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>