Process_CSR: Drop processDBCsrFlagsSet and getActualFlags
This still does not make the code really nice, but I have always been
confused by the existence of both processDBCsrFlagsSet and processDBCsr.
Actually the first is quite unnecessary since the flags are tested again in
processDBCsr. getActualFlags is unnecessary because Framework::error_output
can well be called directly where the error occurs.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>