Browser-CSR: Fix clearing problem of fieldset in webkit-based browsers
The fieldset cannot be "cleared" because then it will not only be cleared
relative to the floating notification message, but also relative to the
floating Confusa menu (and thus appear under the menu). In Firefox it was
sufficient to add a "<br />" to the header of the page, which automatically
cleared the fieldset relative to the notification message. Webkit
disagrees, because to webkit the notification message has a height of 3em
and <br /> only moves the fieldset 1 em down. Thus, a padding-top of 3em
was added to a div wrapping the fieldset.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>