Menu template: Fix the dimensioning for WebKit
WebKit has the somewhat annoying property to automatically insert quite
generous paddings/margins into the styling of <ul>, <p> and <h*> elements.
This resulted in an ugly menu and site contents that were teared apart. By
dictating all margins and paddings for headings and uls, the behaviour of
WebKit should be fixed. Also, specifiy the min-height of the content in em
to excatly match the size of the menu and not letting the menu stick out
below the site div.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>