Subscribers: try to get map from db if subscriber has valid entry
updateFromDB does not try to get the map, it only synchronizes the
subscriber-details from the database. Thus, if this is successful, we
should try to retrieve the map as well.

The getMap() only returns the cached map (to lessen the database load),
and we need to call retrieveMap() to actually get the map.

Signed-off-by: Henrik Austad <henrik@austad.us>