Download-cert: Limit list of certificates to configurable number of days
But give the user a link "Show all" to view all her certificates. This
should have positive effects on the loading time of "My certificates".
The configuration setting for the displayed history defaults to 7 days.

Most often, users will only need a reasonable history of certificates,
for instance 7 days or 3 days. Thus we optimize for the most frequent and
most probable cases while still presenting all options to the user.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>