Download-cert: Force re-auth also with AJAX "Inspect"
Until now, the AJAX code would just wait for status code 200 and then
display whatever code it would receive from the server side. This is
not a good approach - since Confusa will just display the simplesamlphp
login mask in the certificate section when the session times out
(JavaScript calls may not be made to 3rd party domains btw., but simplesaml
usually runs on the same domain as Confusa).

A friendly string can be used to indicate success, in this case the HTML is
just prefixed with Success:

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>