XML_Client: add simple client for authN via SSL + X.509
This is a *very* simple client for SSLv3+X.509 auth.

No error-checking is performed, nor is any parsing or validating of the
returned XML-message performed. The get_list works properly, the
revoke_list does not work yet.

This is intended as a library-function to ease the usage of the RI.

Signed-off-by: Henrik Austad <henrik@austad.us>