Attribute-mapping: Create Subscriber::saveMap() and call from attributes
We already have NREN::saveMap() and hence why not do the same in
Subscriber, especially given that Subscriber has all the information it
needs for updating the map, such as the NREN-Map, it's own map etc.

This even helps save some useless DB-connections and Subscriber can call
retrieveMap() on itself upon an update and thus immediately decorate itself
with the new map-values. This is necessary because otherwise the old map
values still show in the attribute-mapping interface even after change them
until the page is reloaded again.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>