stylist: keep changed notification mail when sending test to own email
When an NREN-admin updates the notification mail-template sent out when
a new certificate has been generated, the changes were lost fromt he
form when the page was updated.

This happened for when a test-mail was generated to inspect the
settings, or when it was saved (the file was stored properly, but the
content was not retrieved adequately, and thus the form missed the new
content).

The reason for this was because the content was assigned blindly using
the stored values without testing for uploaded content_mail.

This change stores the uploaded and changed template and assigns this to
the field if it differs from the stored value thus retaining the values
between "debugging-cycles" and normal updates.

Signed-off-by: Henrik Austad <henrik@austad.us>