Framework: force SSL except in debug-mode.
Confusa is a 'critical' application, and it should *never*, *ever*
accept connections on other than HTTPS when in production-mode.

This change will add test to the very start of Framework, and if HTTPS
is not enabled, it will either exit with a very clear warning, or print
an ugly warning if the portal is placed in debug-mode.

Signed-off-by: Henrik Austad <henrik@austad.us>