Translator: Adapt template decoration to JSON-based dictionaries
As simplesamlphp moved to JSON-based dictionaries, we are using them as
well. Parsing JSON works a bit differently to just including PHP-source
file, but the needed adaptations are not too big. Most additional code
comes from the fact that there are now two files: One with definitions of
the tags and one with the translations of the tags.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>