person: adapt person to handle users with more than one email.
The email-field is a multi-value field, and thus we can expect users
with more than one address. Previously, Confusa would only use the first
address in the lsit, discarding all the rest.

This change will store *all* addresses provided, and extend the getEmail
to return a specified index. The function is backwards-compatible (index
defaults to 0, the old approach).

Signed-off-by: Henrik Austad <henrik@austad.us>