Translator: Move back to content_page
It was not particularly good design to have Translator as a static member
of Framework that was initialized in the constructor, since the static
translator would not work with an uninitialized framework. Since
translator needed non-static information (the person), it was moved back
to content_page, where Framework can access it as well through its
content_page member.

Guessing the best language is deferred until pre_process of every page so
messages output before that (e.g. in the Framework) require a page refresh
before displaying in the correct_language, when the language is changed.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>