index.tpl: make sure subscriber is set before running about_you.
When we display info about a person, we need information about the
user, subscriber and NREN. If the subscriber is not set, the user cannot
use the portal, and a message should be displayed. Furthermore, since
subscriber er is not set, it does not make sense to try to display
information about it.

This change will add a if-else switch around the page, and only when the
person is authenticated and the subscriber is configured into he portal
will about-you info be displayed.

Signed-off-by: Henrik Austad <henrik@austad.us>