Framework: test if smarty cache and compile dir is available.
If the directories are not present and/or writable, the framework would
fail silently. No error-messages would be displayed (as the
template-engine is unable to run) and no direct log-messages would be
displayed.

This change will look for the directories and test if they are
writable. If not, a short message is logged with which folders to create
and enable for writing to the webserver.

Signed-off-by: Henrik Austad <henrik@austad.us>