add_subscriber: set default state in drop-down to 'unsubscribed'
When a new subscriber is added, in most cases an NREN-admin will do the
work in the portal when starting the paperwork. As long as the
subscriber remains 'unsubscribed', it is safe to add the subscriber.

To avoid that new subscriber is added and all its users are given full
access before the paperwork is finalized, the default should be the
safest available option (which is 'unsubscribed').

Signed-off-by: Henrik Austad <henrik@austad.us>