attributes: if the subscriber is not set, the epodn should be set to ""
This happens when the NREN-map is not set, and no subscriber is created
for the current user. Then the framework will bomb up as it tries to get
the subscriber even if it does not exists.

A simple fix is to just set epodn to empty as no subscriber will have
this DN, but Subscriber will handle that when trying to find it. Thus,
we will have a Subscriber-object, but an empty shell.

Signed-off-by: Henrik Austad <henrik@austad.us>