Stylist: Allow for multiple, position-annotated logo files
New logos can be in the top-left, top-center, top-right, background,
bottom-left, bottom-center and bottom-right positions of Confusa.

Those positions are hardcoded, since doing it even more generically is
quite impossible given the limitations of HTML, CSS and the rendering
behaviour of current browsers. When uploading an image, the
intended position is passed in a POST parameter.

This also entails that logos must be removeable, since NREN-admins may not
want to use all of those positions. Thus, a method deleteLogo has been
added, along with the processing of the associated POST parameter.

Finally, since the dimensions should be configurable, Stylist exports the
width-dimensions of the images from the custom CSS to the template (the
height is not restricted but obviously should be proportional to the width).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>