Stylist edit_logo: Change user interface for multiple logo upload
Here a lot of JavaScript is used. Since seven logos are quite many, I
wanted to make the update process somewhat intuitive (not sure if that goal
has been reached), that's why the logos are presented using the same geometry
as in the portal itself. A click on the logo that is to be used causes a few
DOM manipulations to change the UI and fill out form values. Thus, the
right POST parameters concerning the logo's position can be send without
prior data exchange between PHP script and UI.

The page currently works only with JavaScript. The reasoning was that it is
2010 now, really all modern browsers support it and it applies only to the
admin view, updated by a handful of individuals. And a lynx or w3m user is
probably not going to update the graphical logo site.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>