Framework: Use positioned header/footer logos and background if present
The framework just decorates the template as usual, now with up to seven
logos instead of one. The site-template actually dislays them in the right
position. The background image is positioned above the
rounded-corner-div using the respective stylesheet-attribute. This will
actually remove the rounded corners from Confusa. That's the only way to
make it look smooth because with the corners will otherwise always
interfere with the background image.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>