Session: change the name of the session to match SimpleSAMLphp
When we use the session, we must be careful not to destroy it before
simeplsamlphp has finished using it. The translator kicks in before some
of the authentication-steps has finished, and by destroying the session
before this is done, it becomes impossible to authenticate a user
properly.

This change removes the old overwriting of the session, and moves the
session-name from withing the session-handler and into the
constants-collection to make it configurable, but not to clutter the
confusa_config.

Signed-off-by: Henrik Austad <henrik@austad.us>