index: show all emails registred for the user
When we show the about-you part at index, we should show the entire list
of  email-addresses we have available.

This will also add the list with a small icon and a mailto-tag so the
user can test the validity (and where the email ends up) by simply
clicking at it.

Signed-off-by: Henrik Austad <henrik@austad.us>