process_csr: add the email-form to all forms where CSRs are uploaded.
When a user is about to upload a CSR to the server, a list of available
email-addresses should be made available for the user to pick from. This
list should also be dependent upon how the NREN-admin(s) have configured
the portal.

If the portal is configured not to include any email-addresses, nothing
will be displayed, otherwise a radio-list or check-list depending on the
mode

Signed-off-by: Henrik Austad <henrik@austad.us>