I18N: Change the order of determining the default language in Framework
First, instead of the session a cookie is used to store the preferred
language, since that is more stable given the authN redirects and even
persists between user sessions if that is desired by the user.

Introduce a further level of "best" language determination hierarchy - the
functionally set one dominates over everything else. That
translator->setLanguage() function is called in the Framework if there is a
GET present. Thusly, framework error messages (and other messages) are
translated correctly even if the Translator is called before
CP::pre_process, without needing a page refresh.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>