confusa_auth: only add valid entitlements to person.
A user may have multiple entitlements, including multiple in the same
urn-namespace. This is most likely users that have attributes for both
personal and eScience.

When a user sees a list of several attributes, seemingly at random, or
meaning the same, the can (understandably) become confused. This should
simply not happen.

This fix will test each entitlement-attribute that is part of the
namespace to see if it is a valid entitlement for the current
portal. Only if it matches *all* aspects, will it be allowed through.

Signed-off-by: Henrik Austad <henrik@austad.us>