Framework: compose more instructive error-message if entitlement is not set.
There are bound to be users without the required entitlement. These
users should be given a clear warning and directions to where it can be
rectified.

This change splits the entitlement_unset warning into several pieces and
adds reference to the required attribute, the email-address of the
helpdesk as well as webpages.

This will close ticket #217.

Signed-off-by: Henrik Austad <henrik@austad.us>