Process_csr: add javascript to test if UAP has been ticked
In order to make the life a bit simpler for the user, we test to see if
the box for approved UAP has been ticked before sending in the form. If
the box has *not* been ticked, Confusa will reject the request, and the
user must start all over again.

By notifying the user immediately, it becomes clearer, and we remove
several warnings/errors in the system.

Signed-off-by: Henrik Austad <henrik@austad.us>