CA_Comodo: Use the new Confusa_Session for storing the cache
If the simplesamlphp session tied to Person is used, then the session is
always changed from the Confusa session to the simplesamlphp-session and
vice versa. This leads to data loss in the session. It is much better if
only the Confusa session is used for information not directly related to
the simplesamlphp authN process.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>