Confusa_Session: include value of $_SESSION in testSession()
When we test the session, we shouldn't only concern ourselves with the
name, but also if the session is null or not.

The reason for this is as follows: if the session has not started, and
you try to either read from or write to it, PHP will create a *new*
session and drop the existing session.

This will add the is_null test to session, thereby triggering a
session_start if it has not been started yet.

Signed-off-by: Henrik Austad <henrik@austad.us>