messages: add unsubscribed-info to dictionary.
When a user access the page, and the home organization (subscriber) has
not yet subscribed, the user must be given a warning in his/her own
language.

This adds a very rudimentary and *very* simple message setup with 2
paragraphs of text and a header. Normally, a user should not see this,
and if so, the subscriber is most likely in the process of subscribing
to the service.

Signed-off-by: Henrik Austad <henrik@austad.us>