confusa_handler: mask error-code to error-bitmask to remove excess warnings.
Confusa has a custom PHP error logger which is very nice.
One problem with it though is that it completely bypasses the PHP error handler,
and thus the minimum setting for error reporting.as defined in php.ini.

Problem with this is that e.g. E_STRICT, which is triggered quite a
lot by MDB and PEAR, swamps the logs with no way to turn it off.

By masking the error-code to the error-reporting bitmask, this can be resolved.

Signed-off-by: Henrik Austad <henrik@austad.us>