process_csr: make sure the uap is set when approving the browser-csr
This is safe to do, as we know that the user must have gone through the
webform and therefore the uap-box. This is an extra layer of
security. It should not be possible to get to this point without
approving the UAP, but "It should not be possible" can pose as a very
interesting challenge, thus we play it safe :-)

Signed-off-by: Henrik Austad <henrik@austad.us>