CA: Make sure the CA-config-switch is valid.
If the config is malformed, we should log it so the operators can figure
it out. This fix will return null and add a log-entry. However, this
will also break most of the portal in a pretty spectacular way.

Hopefully, the people operating the portal will test it before releasing
it to the public, and they *will* notice this problem.

Signed-off-by: Henrik Austad <henrik@austad.us>